Rework interpreter
After coming back to this after leaving it for a few months, I found quite a few things I wanted to change about the interpreter, and ending up more-or-less rewriting it. The main thing that was changed was that I added a few functions for manipulating the data structures that the interpreter uses, although quite a few miscellaneous changes are present.